+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.3.1
+
+- Add documentation to existing exports
+
+## 0.2.0
+
+- Add stoyco subscription circular image with info and fonts
+
+## 0.1.0
+
+- Add fonts and color utils. Add subscription search bar widget
+
 ## 0.0.1
 
-* Init library
+- TODO: Describe initial release.